refactor: replace custom Card with primitive
